requirements.txt
playwright>=1.40.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
